Main validator in separate file
